Removing duplicates: Route refactoring 2
