+++ src/backends/envoy_xds_backend/mod.rs
@@ -2,5 +2,6 @@ mod converters;
 mod envoy_deployer;
 mod model;
 mod resources;
+mod route_converters;
 mod server;
 pub use envoy_deployer::EnvoyDeployerChannelHandlerService;
