+++ src/common/route/mod.rs
@@ -3,10 +3,11 @@ pub mod http_route;
 
 use std::cmp;
 
-use gateway_api::common_types::{HTTPHeader, HeaderModifier, RouteRef};
+use gateway_api::common_types::{HTTPHeader, HeaderModifier, MatchingHeaders, RouteRef};
 pub use grpc_route::GRPCEffectiveRoutingRule;
 pub use http_route::HTTPEffectiveRoutingRule;
 use thiserror::Error;
+use typed_builder::TypedBuilder;
 
 use super::{Backend, BackendServiceConfig, ResourceKey, DEFAULT_NAMESPACE_NAME, DEFAULT_ROUTE_HOSTNAME};
 use crate::common::route::{grpc_route::GRPCRoutingConfiguration, http_route::HTTPRoutingConfiguration};
@@ -178,4 +179,45 @@ pub struct FilterHeaders {
     pub add: Vec<HTTPHeader>,
     pub remove: Vec<String>,
     pub set: Vec<HTTPHeader>,
+}
+
+struct Comparator<'a, T> {
+    this: Option<&'a Vec<T>>,
+    other: Option<&'a Vec<T>>,
+}
+
+impl<T> Comparator<'_, T> {
+    pub fn compare(self) -> std::cmp::Ordering {
+        match (self.this, self.other) {
+            (None, None) => std::cmp::Ordering::Equal,
+            (None, Some(_)) => std::cmp::Ordering::Greater,
+            (Some(_), None) => std::cmp::Ordering::Less,
+            (Some(this_headers), Some(other_headers)) => other_headers.len().cmp(&this_headers.len()),
+        }
+    }
+}
+
+#[derive(TypedBuilder)]
+struct HeaderComparator<'a> {
+    this: Option<&'a Vec<MatchingHeaders>>,
+    other: Option<&'a Vec<MatchingHeaders>>,
+}
+
+impl HeaderComparator<'_> {
+    pub fn compare_headers(self) -> std::cmp::Ordering {
+        let comp = Comparator { this: self.this, other: self.other };
+        comp.compare()
+    }
+}
+
+#[derive(TypedBuilder)]
+struct QueryComparator<'a> {
+    this: Option<&'a Vec<MatchingHeaders>>,
+    other: Option<&'a Vec<MatchingHeaders>>,
+}
+impl QueryComparator<'_> {
+    pub fn compare_queries(self) -> std::cmp::Ordering {
+        let comp = Comparator { this: self.this, other: self.other };
+        comp.compare()
+    }
 }
